requirements.txt
django
djangorestframework
requests
django-cors-headers
psycopg2-binary
channels
python-dotenv
channels-redis

